+++ modules/function_handle/src/include/nlsFunction_handle_exports.h
@@ -1,0 +1,22 @@
+//=============================================================================
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#pragma once
+//=============================================================================
+#ifdef _MSC_VER
+#ifdef NLSFUNCTION_HANDLE_EXPORTS
+#define NLSFUNCTION_HANDLE_IMPEXP __declspec(dllexport)
+#else
+#define NLSFUNCTION_HANDLE_IMPEXP __declspec(dllimport)
+#endif
+#else
+#define NLSFUNCTION_HANDLE_IMPEXP __attribute__((visibility("default")))
+#endif
+//=============================================================================